+++ storehaus-core/src/main/scala/com/twitter/storehaus/IndexedSeqReadableStore.scala
@@ -18,6 +18,7 @@ package com.twitter.storehaus
 
 import com.twitter.util.Future
 
+/** Simple wrapper on IndexedSeq[V] to make it accessible as a ReadableStore */
 class IndexedSeqReadableStore[+V](iseq: IndexedSeq[V]) extends ReadableStore[Int, V] {
   override def get(idx: Int) = if (idx >= 0 && idx < iseq.size) {
     Future.value(Some(iseq(idx)))