+++ storehaus-core/src/test/scala/com/twitter/storehaus/ReplicatedStoreProperties.scala
@@ -19,8 +19,6 @@ package com.twitter.storehaus
 import org.scalacheck.Properties
 
 object ReplicatedStoreProperties extends Properties("ReplicatedStore") {
-  import StoreProperties.storeTest
-
   property("ReplicatedStore test") =
-    storeTest(new ReplicatedStore(Stream.continually(new JMapStore[String, Int]).take(100).toSeq))
+    StoreProperties.storeTest(Store.first(Stream.continually(new JMapStore[String, Int]).take(100).toSeq))
 }